Reuse Back Corner player in admin
